requirements.txt
fastapi
uvicorn
transformers
pydantic
pytest - # for testing
httpx - # for making my HTTP requests in tests
pyyaml - # for my YAML configuration